添加大量功能.
1.添加jquery UI日期控件.
2.添加POJO.
3.完成服务申请下的4个子申请的添加代码.
4.页面未做验证.
5.更新sql脚本.

+++ src/main/java/com/sobey/mvc/dao/account/GroupDao.java
@@ -2,13 +2,13 @@ package com.sobey.mvc.dao.account;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import com.sobey.mvc.entity.account.Group;
+import com.sobey.mvc.entity.Group;
 
 /**
  * 权限组对象的Dao interface.
  * 
  */
 
-public interface GroupDao extends PagingAndSortingRepository<Group, Long>, GroupDaoCustom {
+public interface GroupDao extends PagingAndSortingRepository<Group, Integer>, GroupDaoCustom {
 	Group findByName(String name);
 }
