+++ src/main/java/com/sobey/mvc/dao/account/GroupDaoCustom.java
@@ -8,6 +8,6 @@ public interface GroupDaoCustom {
 	/**
 	 * 因为Group中没有建立与User的关联,因此需要以较低效率的方式进行删除User与Group的多对多中间表中的数据.
 	 */
-	void deleteWithReference(Long id);
+	void deleteWithReference(Integer id);
 
 }
